tests/requirements.txt
pytest
pytest-describe
pytest-pythonpath
pytest-django
tox >= 1.8
